requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
google-generativeai==0.3.2
langchain==0.1.0
langchain-community==0.0.13
langchain-pinecone==0.0.3
sentence-transformers==2.2.2
pinecone-client==3.0.0
huggingface-hub==0.19.4